+++ packages/terra-framework-docs/src/terra-dev-site/doc/table/Examples/TableRowSelection.jsx
@@ -88,11 +88,6 @@ const TableRowSelection = () => {
     setRowData(newRowData);
   }, [rowData]);
 
-  const onRowSelectionHeaderSelect = useCallback(() => {
-    // eslint-disable-next-line no-alert
-    alert('Row Selection Header Clicked');
-  }, []);
-
   return (
     <React.Fragment>
       <div>
@@ -113,7 +108,6 @@ const TableRowSelection = () => {
         ariaLabel="Table with Row Selections"
         rowSelectionMode={hasSelectableRows ? 'multiple' : undefined} // Prop to turn row selection mode on/off
         onRowSelect={onRowSelect} // For row selection, consumer must provide a callback that the Worklist Data Grid will call when the user selects one or more rows.
-        onRowSelectionHeaderSelect={onRowSelectionHeaderSelect}
       />
     </React.Fragment>
   );